+++ src/constants/countries.js
@@ -1,0 +1,11 @@
+export const COUNTRIES = {
+  'United Kingdom': { currency: 'GBP', symbol: '£' },
+  'United States': { currency: 'USD', symbol: '$' },
+  'Ukraine': { currency: 'UAH', symbol: '₴' },
+  'European Union': { currency: 'EUR', symbol: '€' },
+  'Japan': { currency: 'JPY', symbol: '¥' },
+  'Canada': { currency: 'CAD', symbol: '$' },
+  'Australia': { currency: 'AUD', symbol: '$' },
+  'Poland': { currency: 'PLN', symbol: 'zł' },
+  'Latvia': { currency: 'EUR', symbol: '€' },
+};
