+++ src/constants/languages.js
@@ -1,0 +1,2 @@
+export const LANGUAGES = ['English', 'Russian'];
+export const PREDEFINED_MODELS = ['gemini-2.0-flash', 'gemini-2.5-flash-preview-05-20'];
