+++ src/constants/translations.js
@@ -1,0 +1,124 @@
+export const translations = {
+    English: {
+        travelPlanner: "Travel Planner",
+        newTrip: "New Trip",
+        myTrips: "My Trips",
+        adventuresAwait: "Your adventures await!",
+        noTripsPlanned: "No trips planned yet.",
+        clickNewTrip: "Click \"New Trip\" to start your next journey!",
+        backToAllTrips: "Back to all trips",
+        backToTripOverview: "Back to trip overview",
+        loadingPlanner: "Loading Planner...",
+        loadingTrip: "Loading trip...",
+        loadingSection: "Loading section...",
+        settings: "Settings",
+        done: "Done",
+        cancel: "Cancel",
+        create: "Create",
+        save: "Save",
+        delete: "Delete",
+        confirm: "Confirm",
+        deleteTripTitle: "Delete Trip?",
+        deleteTripMessage: "Are you sure you want to permanently delete this trip?",
+        language: "Language",
+        myLocation: "My Location",
+        geminiApiKey: "Gemini API Key",
+        apiKeyHint: "Get your key from Google AI Studio.",
+        apiKeyPlaceholder: "Enter key for AI features",
+        createANewTrip: "Create a New Trip",
+        editTrip: "Edit Trip",
+        tripNamePlaceholder: "Trip Name (e.g., Paris Getaway)",
+        locationPlaceholder: "Location",
+        tripColor: "Trip Color",
+        thingsToTake: "Things to Take",
+        placesToVisit: "Places to Visit",
+        usefulLinks: "Useful Links",
+        expenses: "Expenses",
+        thingsToTakePlaceholder: "e.g., Passport",
+        placesToVisitPlaceholder: "e.g., Eiffel Tower",
+        usefulLinksPlaceholder: "e.g., Museum tickets",
+        expensesPlaceholder: "e.g., Flights",
+        aiSuggest: "AI Suggest",
+        generating: "Generating...",
+        descriptionOptional: "Description (optional)",
+        ai: "AI",
+        aiSettings: "AI Settings",
+        model: "Model",
+        modelOther: "Other...",
+        modelCustomPlaceholder: "Enter custom model name",
+        aiDescriptionError: "Failed to generate description. Error:",
+        aiSuggestError: "Failed to get AI suggestions. Error:",
+        enterApiKey: "Please enter your Gemini API key in Settings.",
+        unexpectedFormat: "AI suggestions were in an unexpected format.",
+        enterItemName: "Please enter an item name first.",
+        nothingYet: "Nothing here yet...",
+        addSection: "Add Section",
+        newSectionName: "New Section Name",
+        sectionNamePlaceholder: "e.g., Souvenir Ideas",
+        sectionNameError: "Section name cannot be empty or a duplicate.",
+        deleteSelected: "Delete Selected",
+        deleteSelectedTripsTitle: "Delete Selected Trips?",
+        deleteSelectedTripsMessage: "Are you sure you want to permanently delete the selected trips?",
+    },
+    Russian: {
+        travelPlanner: "Планировщик поездок",
+        newTrip: "Новая поездка",
+        myTrips: "Мои поездки",
+        adventuresAwait: "Ваши приключения ждут!",
+        noTripsPlanned: "Пока нет запланированных поездок.",
+        clickNewTrip: "Нажмите «Новая поездка», чтобы начать новое путешествие!",
+        backToAllTrips: "Назад ко всем поездкам",
+        backToTripOverview: "Назад к обзору поездки",
+        loadingPlanner: "Загрузка планировщика...",
+        loadingTrip: "Загрузка поездки...",
+        loadingSection: "Загрузка раздела...",
+        settings: "Настройки",
+        done: "Готово",
+        cancel: "Отмена",
+        create: "Создать",
+        save: "Сохранить",
+        delete: "Удалить",
+        confirm: "Подтвердить",
+        deleteTripTitle: "Удалить поездку?",
+        deleteTripMessage: "Вы уверены, что хотите навсегда удалить эту поездку?",
+        language: "Язык",
+        myLocation: "Мое местоположение",
+        geminiApiKey: "API-ключ Gemini",
+        apiKeyHint: "Получите ключ в Google AI Studio.",
+        apiKeyPlaceholder: "Введите ключ для функций ИИ",
+        createANewTrip: "Создать новую поездку",
+        editTrip: "Редактировать поездку",
+        tripNamePlaceholder: "Название поездки (например, Поездка в Париж)",
+        locationPlaceholder: "Местоположение",
+        tripColor: "Цвет поездки",
+        thingsToTake: "Что взять с собой",
+        placesToVisit: "Места для посещения",
+        usefulLinks: "Полезные ссылки",
+        expenses: "Расходы",
+        thingsToTakePlaceholder: "например, Паспорт",
+        placesToVisitPlaceholder: "например, Эйфелева башня",
+        usefulLinksPlaceholder: "например, Билеты в музей",
+        expensesPlaceholder: "например, Авиабилеты",
+        aiSuggest: "ИИ-подсказка",
+        generating: "Генерация...",
+        descriptionOptional: "Описание (необязательно)",
+        ai: "ИИ",
+        aiSettings: "Настройки ИИ",
+        model: "Модель",
+        modelOther: "Другая...",
+        modelCustomPlaceholder: "Введите свое название модели",
+        aiDescriptionError: "Не удалось сгенерировать описание. Ошибка:",
+        aiSuggestError: "Не удалось получить подсказки ИИ. Ошибка:",
+        enterApiKey: "Пожалуйста, введите ваш API-ключ Gemini в настройках.",
+        unexpectedFormat: "Подсказки ИИ были в неожиданном формате.",
+        enterItemName: "Пожалуйста, сначала введите название элемента.",
+        nothingYet: "Здесь пока ничего нет...",
+        addSection: "Добавить раздел",
+        newSectionName: "Название нового раздела",
+        sectionNamePlaceholder: "например, Идеи для сувениров",
+        sectionNameError: "Название раздела не может быть пустым или дублироваться.",
+        deleteSelected: "Удалить выбранные",
+        deleteSelectedTripsTitle: "Удалить выбранные поездки?",
+        deleteSelectedTripsMessage: "Вы уверены, что хотите навсегда удалить выбранные поездки?",
+    },
+};
