+++ src/hooks/useAutosizeTextArea.js
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+export const useAutosizeTextArea = (textAreaRef, value) => {
+  useEffect(() => {
+    if (textAreaRef.current) {
+      textAreaRef.current.style.height = '0px';
+      const scrollHeight = textAreaRef.current.scrollHeight;
+      textAreaRef.current.style.height = scrollHeight + 'px';
+    }
+  }, [textAreaRef, value]);
+};
